Use GetText::Tools::Task introduced since gettext 3.0.0
This commit is for #101.

This commit requires gettext 3.0.1 (it is not released yet) or
later. Older gettext doesn't support. (Is it OK?)

This commit also disables lazy gettext gem require. Because we don't
use GetText::Tools::Task after requiring gettext gem. (Is it OK?)

We can add shim code to support both gettext 3.0.1 or later and old
versions. We try to require `gettext/tools/task`. If we succeed it, we
assume that we have gettext 3.0.1 or later. We use this commit's
code. If we fail it, we assume that we have old gettext. We use the
previous code.