Switch out haml and slim parsers to use GetText::RubyParser
We now just parse the haml or slim to ruby code and then hand it off to the GetText::RubyParser to actually extract the strings into the PO format.

This adds support for translations with context (`p_("My Context", "id_key")`) and removes the custom ruby extractor entirely.

It also fixes the 'pending' spec for slim to parse ruby 1.9 syntax.

Fixes: https://github.com/grosser/gettext_i18n_rails/issues/199